requirements.txt
pandas
plotly
streamlit==1.35.0
pandas-datareader
matplotlib==3.5.2
streamlit_folium
geopandas
st-pages
transformers
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
folium
jinja2
branca
numpy
seaborn
plotly
tokenizers
nltk
wordcloud
pillow